Cleaned up index.tsx
Moved some code around to clean up index.tsx
Automatically switch to Glitch.com server when on production
